fix: drop duplicate "CarrierType", fallback to CarrierCode
